data-engineering/etl_pipeline_enriches_json/requirements.txt
apache-airflow==2.8.1
pandas
pytest
requests
tenacity